parser/chart_detector.py
"""
Chart Detector Module
======================

Provides basic detection of charts and diagrams via image analysis or
placeholder logic. Designed for extension with OCR or ML-based analysis.
"""

import logging
from typing import Dict, Any
import fitz  # PyMuPDF

from utils.config import Config


class ChartDetector:
    """
    Chart and diagram detection utilities.
    """
    def __init__(self, config: Config):
        self.config = config
        self.logger = logging.getLogger(__name__)

    def detect_charts(self, fitz_page: fitz.Page) -> Dict[str, Any]:
        """
        Detect embedded chart images within a page.

        Returns placeholder classifications for each chart found.
        """
        detected = []
        for img in fitz_page.get_images():
            try:
                bbox = fitz_page.get_image_bbox(img[0])
                detected.append({
                    'type': 'chart',
                    'bbox': {
                        'top': bbox.y0,
                        'left': bbox.x0,
                        'width': bbox.width,
                        'height': bbox.height
                    },
                    'description': 'Detected chart image',
                    'confidence': 0.7
                })
            except Exception as e:
                self.logger.debug(f"Error detecting chart image: {e}")
        return detected
